Changed contextId to long, implemented method of UserGroupHandle.

+++ openexchange-test/src/com/openexchange/groupware/contexts/TestContextImpl.java
@@ -19,7 +19,7 @@ public class TestContextImpl implements Context {
     /**
      * {@inheritDoc}
      */
-    public int getContextId() {
+    public long getContextId() {
         return contextId;
     }
